move stabilize cli to register, startin;5D;5D;5D;5D;5D;5Dg refactor to one cli
